Added Section1 and the Front Logo
- Updated Base/Base.scad
- Added Base/Section1.{stl,parts.stl}
- Added Base/Section1/*
- Updated include/parts/logo.scad
  Included a little tagline (may not be permanent).
  I suppose experiementing with social reactions to
  arrangements of symbols that convey a fairly accurate message.

Other random stuffness:

So it seems that Makers Faire SF will not be happening?
I'm a little unclear, many of the people I'm "working" with are
not very good about making sure things are organized; traditionally
that's been my boat.

Seems that socially I've fallen out and the imperitive reactions
are only pushing people away or something. I keep getting told
that <person> is upset that I'm not producing... .

So to respond I do this... if this is not producing then...
I mean I'm gonna reinvent the wheel; sorry if that bothers you.

I might just do it using magnetorquer technology or some other sci-fi bull.
Because I mean, I'm bored: so let's go already! It's kinda rough to be
told to communicate and lead this when I'm so disconnected.